finilized and tested custom training
