+++ stacc/__init__.py
@@ -1,2 +1,7 @@
 from .__version__ import __version__
-from prediction import run_counting
+from .prediction import run_counting
+from .dataset import StaccImageCollectionDataset
+from .dataset import StaccDataLoader
+from .util import load_config
+from .util import get_device
+from .training.stacc_training import run_stacc_training
