--- stacc/training/__init__.py
@@ -1,2 +1,0 @@
-from .stacc_training import run_stacc_training
-from .utils import StaccDataLoader
